+++ node/index.ts
@@ -9,6 +9,9 @@ import { LRUCache, Service } from '@vtex/api'
 import { Clients } from './clients'
 import { resolvers } from './resolvers'
 import { schemaDirectives } from './resolvers/directives'
+import { appDependenciesUpdatedEvent } from './events/appDependenciesUpdatedEvent'
+import { kubeRouterAuditEvent } from './events/kubeRouterAuditEvent'
+import { appAllEvents } from './events/appAllEvents'
 
 const TIMEOUT_MS = 4000
 
@@ -51,4 +54,9 @@ export default new Service<Clients, RecorderState, ParamsContext>({
     schemaDirectives,
   },
   routes: resolvers.Routes,
+  events: {
+    appDependenciesUpdatedEvent,
+    kubeRouterAuditEvent,
+    appAllEvents,
+  },
 })